fastapi/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6
# pydantic==2.5.0
# python-dotenv==1.0.0
# langchain==0.1.0
# langchain-openai==0.0.2
# langchain-community==0.0.10
# chromadb==0.4.18
# pypdf==3.17.4


fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1

langchain==0.2.16
langchain-openai==0.1.22
langchain-community==0.2.16
langsmith==0.1.119

chromadb==0.5.4
pypdf==4.3.1
